program-runtime: port over remaining cpi tests
